+++ st2actions/st2actions/runners/fabricrunner.py
@@ -21,12 +21,12 @@ from oslo.config import cfg
 import six
 
 from st2actions.runners import ActionRunner
+from st2actions.runners import ShellRunnerMixin
 from st2common import log as logging
 from st2common.exceptions.actionrunner import ActionRunnerPreRunError
 from st2common.exceptions.fabricrunner import FabricExecutionFailureException
 from st2common.constants.action import ACTIONEXEC_STATUS_SUCCEEDED, ACTIONEXEC_STATUS_FAILED
 from st2common.models.system.action import (FabricRemoteAction, FabricRemoteScriptAction)
-import st2common.util.action_db as action_utils
 
 # Replace with container call to get logger.
 LOG = logging.getLogger(__name__)
@@ -67,10 +67,9 @@ def get_runner():
     return FabricRunner(str(uuid.uuid4()))
 
 
-class FabricRunner(ActionRunner):
-    def __init__(self, id):
-        super(FabricRunner, self).__init__()
-        self._runner_id = id
+class FabricRunner(ActionRunner, ShellRunnerMixin):
+    def __init__(self, runner_id):
+        super(FabricRunner, self).__init__(runner_id=runner_id)
         self._hosts = None
         self._parallel = True
         self._sudo = False
@@ -96,7 +95,7 @@ class FabricRunner(ActionRunner):
         self._timeout = self.runner_parameters.get(RUNNER_TIMEOUT, DEFAULT_ACTION_TIMEOUT)
 
         LOG.info('[FabricRunner="%s", actionexec_id="%s"] Finished pre_run.',
-                 self._runner_id, self.action_execution_id)
+                 self.runner_id, self.action_execution_id)
 
     def run(self, action_parameters):
         LOG.debug('    action_parameters = %s', action_parameters)
@@ -115,7 +114,7 @@ class FabricRunner(ActionRunner):
 
     def _run(self, remote_action):
         LOG.info('Executing action via FabricRunner :%s for user: %s.',
-                 self._runner_id, remote_action.get_on_behalf_user())
+                 self.runner_id, remote_action.get_on_behalf_user())
         LOG.info(('[Action info] name: %s, Id: %s, command: %s, on behalf user: %s, '
                   'actual user: %s, sudo: %s'),
                  remote_action.name, remote_action.id, remote_action.get_command(),
@@ -141,7 +140,7 @@ class FabricRunner(ActionRunner):
     def _get_fabric_remote_script_action(self, action_parameters):
         script_local_path_abs = self.entry_point
         pos_args, named_args = self._get_script_args(action_parameters)
-        named_args = self._transform_pos_args(named_args)
+        named_args = self._transform_named_args(named_args)
         env_vars = self._get_env_vars()
         remote_dir = self.runner_parameters.get(RUNNER_REMOTE_DIR,
                                                 cfg.CONF.ssh_runner.remote_dir)
@@ -160,22 +159,6 @@ class FabricRunner(ActionRunner):
                                         parallel=self._parallel,
                                         sudo=self._sudo,
                                         timeout=self._timeout)
-
-    def _transform_pos_args(self, named_args):
-        if named_args:
-            return {self._kwarg_op + k: v for (k, v) in six.iteritems(named_args)}
-        return None
-
-    def _get_script_args(self, action_parameters):
-        is_script_run_as_cmd = self.runner_parameters.get(RUNNER_COMMAND, None)
-        pos_args = ''
-        named_args = {}
-        if is_script_run_as_cmd:
-            pos_args = self.runner_parameters.get(RUNNER_COMMAND, '')
-            named_args = action_parameters
-        else:
-            pos_args, named_args = action_utils.get_args(action_parameters, self.action)
-        return pos_args, named_args
 
     def _get_env_vars(self):
         return {'st2_auth_token': self.auth_token.token} if self.auth_token else {}