+++ st2actions/tests/test_mistral_v2.py
@@ -33,12 +33,14 @@ from st2tests import DbTestCase
 import st2actions.bootstrap.runnersregistrar as runners_registrar
 from st2actions import worker
 from st2actions.runners.mistral.v2 import MistralRunner
-from st2actions.runners.fabricrunner import FabricRunner
+from st2actions.runners.localrunner import LocalShellRunner
 from st2actions.handlers.mistral import MistralCallbackHandler
 from st2common.transport.publishers import CUDPublisher
 from st2common.services import action as action_service
 from st2common.models.db.action import ActionExecutionDB
-from st2common.constants.action import ACTIONEXEC_STATUS_SUCCEEDED, ACTIONEXEC_STATUS_RUNNING
+from st2common.constants.action import ACTIONEXEC_STATUS_SUCCEEDED
+from st2common.constants.action import ACTIONEXEC_STATUS_RUNNING
+from st2common.constants.action import ACTIONEXEC_STATUS_FAILED
 from st2common.models.api.action import ActionAPI
 from st2common.persistence.action import Action, ActionExecution
 
@@ -56,7 +58,7 @@ def process_create(payload):
         CHAMPION.execute_action(payload)
 
 
-@mock.patch.object(FabricRunner, '_run', mock.MagicMock(return_value={}))
+@mock.patch.object(LocalShellRunner, 'run', mock.MagicMock(return_value={}))
 @mock.patch.object(CUDPublisher, 'publish_create', mock.MagicMock(side_effect=process_create))
 @mock.patch.object(CUDPublisher, 'publish_update', mock.MagicMock(return_value=None))
 class TestMistralRunner(DbTestCase):
@@ -161,7 +163,7 @@ class TestMistralRunner(DbTestCase):
             callback={'source': 'mistral', 'url': 'http://localhost:8989/v2/tasks/12345'})
         execution = action_service.schedule(execution)
         execution = ActionExecution.get_by_id(str(execution.id))
-        self.assertEqual(execution.status, ACTIONEXEC_STATUS_SUCCEEDED)
+        self.assertEqual(execution.status, ACTIONEXEC_STATUS_FAILED)
         requests.request.assert_called_with('PUT', execution.callback['url'],
-                                            data=json.dumps({'state': 'SUCCESS', 'result': '{}'}),
+                                            data=json.dumps({'state': 'ERROR', 'result': 'None'}),
                                             headers={'content-type': 'application/json'})